Avoid exception handling for verifying algorithm validity (#212)
Currently, when a `Symbol` is passed to the `Opt` constructor or the
`algorithm_name` function, the corresponding `Algorithm` object is
obtained via a `Dict` lookup inside of a `try`/`catch`, where the
`catch` permits throwing a more informative error. The use of exception
handling here isn't necessary though; rather than using `getindex` and
catching any resulting exception, we can instead use `get` with a
default value that does not correspond to a valid algorithm and check
for equality with that. (Alternatively we could guard the `getindex`
with a `haskey` but that requires performing the lookup twice.) This
change to using `get` is implemented here.

To avoid duplicating the code that checks a `Symbol` for validity as an
algorithm across multiple functions, I've defined a constructor method
`Algorithm(::Symbol)` that includes this check. This also provides a
pleasant simplification of the methods required, since `Symbol` and
`Integer` inputs can now go through the same `Opt` and `algorithm_name`
methods.